requirements.txt
django
django-crispy-forms
crispy-tailwind
django-tailwind==3.8.0
django-browser-reload==1.12.1
django-widget-tweaks==1.5.0
django-import-export==3.3.7
geopy==2.4.1
